refactor(agents): isolate ACP spawn results
