+++ src/agents/subagents/spawn/acp-spawn.ts
@@ -86,6 +86,11 @@ import {
   resolveConfiguredAcpSubagentTargetIds,
   resolveTargetAcpAgentId,
 } from "./acp-spawn-target.js";
+import {
+  createAcpSpawnFailure,
+  type SpawnAcpMode,
+  type SpawnAcpResult,
+} from "./acp-spawn-result.js";
 import { readParentExecutionIdentity } from "./execution-identity-spawn-context.js";
 import {
   isSubagentEnvelopeSession,
@@ -95,7 +100,6 @@ import { readGatewayRunId } from "./subagent-spawn-gateway.js";
 import { resolveSubagentSpawnOwnership } from "./subagent-spawn-ownership.js";
 import { resolveConfiguredSubagentRunTimeoutSeconds } from "./subagent-spawn-plan.js";
 
-type SpawnAcpMode = "run" | "session";
 type SpawnAcpSandboxMode = "inherit" | "require";
 
 type SpawnAcpParams = {
@@ -140,48 +144,6 @@ type SpawnAcpContext = {
   inheritedToolDenylist?: string[];
 };
 
-const ACP_SPAWN_ERROR_CODES = [
-  "acp_disabled",
-  "requester_session_required",
-  "runtime_policy",
-  "resume_forbidden",
-  "subagent_policy",
-  "thread_required",
-  "target_agent_required",
-  "runtime_agent_mismatch",
-  "agent_forbidden",
-  "cwd_resolution_failed",
-  "thread_binding_invalid",
-  "spawn_failed",
-  "dispatch_failed",
-] as const;
-type SpawnAcpErrorCode = (typeof ACP_SPAWN_ERROR_CODES)[number];
-
-type SpawnAcpResultFields = {
-  childSessionKey?: string;
-  runId?: string;
-  mode?: SpawnAcpMode;
-  runTimeoutSeconds?: number;
-  expectsCompletionMessage?: boolean;
-  inlineDelivery?: boolean;
-  note?: string;
-};
-
-type SpawnAcpAcceptedResult = SpawnAcpResultFields & {
-  status: "accepted";
-  childSessionKey: string;
-  runId: string;
-  mode: SpawnAcpMode;
-};
-
-type SpawnAcpFailedResult = SpawnAcpResultFields & {
-  status: "forbidden" | "error";
-  error: string;
-  errorCode: SpawnAcpErrorCode;
-};
-
-type SpawnAcpResult = SpawnAcpAcceptedResult | SpawnAcpFailedResult;
-
 const ACP_SPAWN_ACCEPTED_NOTE =
   "initial ACP task queued in isolated session; follow-ups continue in the bound thread.";
 const ACP_SPAWN_SESSION_ACCEPTED_NOTE =
@@ -204,22 +166,6 @@ export function resolveAcpSpawnRuntimePolicyError(params: {
     requesterSandboxed: params.requesterSandboxed === true || requesterRuntime.sandboxed,
     sandbox: params.sandbox === "require" ? "require" : "inherit",
   });
-}
-
-function createAcpSpawnFailure(params: {
-  status: "forbidden" | "error";
-  errorCode: SpawnAcpErrorCode;
-  error: string;
-  childSessionKey?: string;
-  runId?: string;
-}): SpawnAcpFailedResult {
-  return {
-    status: params.status,
-    errorCode: params.errorCode,
-    error: params.error,
-    ...(params.childSessionKey ? { childSessionKey: params.childSessionKey } : {}),
-    ...(params.runId ? { runId: params.runId } : {}),
-  };
 }
 
 export { resolveRuntimeCwdForAcpSpawn } from "./acp-spawn-runtime.js";
